requirements.txt
azure-functions
azure-storage-blob
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
requests
Pillow
